+++ example/lib/utils.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+
+showSnack(BuildContext context, String text) {
+  ScaffoldMessenger.of(context).showSnackBar(
+    SnackBar(
+      content: Text(
+        text,
+        style: TextStyle(color: Theme.of(context).colorScheme.surface),
+      ),
+      backgroundColor: Theme.of(context).colorScheme.onSurface,
+      duration: Duration(seconds: 1),
+      showCloseIcon: true,
+      closeIconColor: Theme.of(context).colorScheme.surface,
+    ),
+  );
+}