+++ lib/src/utils/extension.dart
@@ -14,7 +14,8 @@ extension DateTimeExtensions on DateTime {
 
   /// get startOfWeek day depending on the desired starting day of the week
   DateTime startOfWeek(int weekday) {
-    return DateTime(year, month, day - this.weekday % weekday);
+    var dayDifference = (this.weekday - weekday) % 7;
+    return DateTime(year, month, day - dayDifference);
   }
 }